Improve gui and service
