+++ jtop/jtop.py
@@ -563,7 +563,7 @@ class jtop(Thread):
         # -- Engines --
         for group in self.engine:
             for name, engine in self.engine[group].items():
-                stats[name] = engine['curr'] if engine['status'] else 'OFF'
+                stats[name] = engine['curr'] if engine['online'] else 'OFF'
         # -- FAN --
         if self.fan:
             stats['fan'] = self.fan.speed
